Use static config file to define the DatoCMS environment (#26)
# Changes

I came to the conclusion that I like a file better than what we tried to
in #14 . For several reasons:

- Cloudlfare checks out a single commit and not a branch, so it is hard
to get the branch name. We can use an Cloudlfare specific env variable
but then we would create something cloudlfare specific and we would need
to normalise it during development. There are packages that normalise
those env variables so It can be done but it complicates things.
- I had my doubt about using branch names already, especially with the
non supporting characters, but if we look at a developer workflow:
After reviewing a PR with a environment attached to it, you would create
a new fork of the primary environment and run the migration script on
the up to date content. Since you can not rename environments its hard
to keep the branch name and environment in sync, since there probably
was already an outdated environment with the name of the branch.

All of the above vs a single JS file with an export of the active
environment. It works on pr branches at when you merge to main it will
still work since you promote the environment of your feature branch to
the primary one.

What do you think?

---------

Co-authored-by: Jasper Moelker <jasper@voorhoede.nl>