Refactor: allow blocks with just `__typename` in AnyBlock (#259)
Consider the type of AnyBlock: it is a union of all defined fragments.
Such a fragment might look like this:
```ts
export type ActionBlockFragment = {
  __typename: 'ActionBlockRecord',
  id: string,
  items: Array<...>
}
export type ImageBlockFragment = {
  __typename?: 'ImageBlockRecord', // not sure why this one is undefined, I copy/pasted it from `datocms/types.ts`
  id: string
  image: { ... }
};
```

Now one might think passing a subset of `AnyBlock` to `Blocks.astro` as
props would be fine, because that's what the union is for, right?
Unfortunately, we're doing something else every time we're listing
blocks, which is actually ruining this for us, and that is including
`__typename` in the list of blocks. This makes it possible to emit to
the console that we've not implemented a block renderer correctly, but
it also means that these are the types generated for it:

```ts
blocks: Array<
    {
      __typename: 'ActionBlockRecord'
      id: string
      items: Array< ... >
    } | { 
      __typename: 'ImageBlockRecord' 
    } 
```

This means that any block that is not allowed wil still be present as a
`__typename: 'SomethingRecord'` type. And, of course, `{ __typename:
'ImageBlockRecord' }`, is not assignable to type ` { __typename:
'ImageBlockRecord', id: string, image { ... } }`.

# Changes

- Loosen type of AnyBlock so that `{ __typename }` can be passed, even
when the Block is not used in that context
- Use types from graphQL query instead of adding AnyBlock type assertion

# Associated issue

# How to test

1. Open preview link
2. See that page still renders
3. Inspect linting check
4. Verify that there are no issues.

# Checklist

- [x] I have performed a self-review of my own code
- [x] I have made sure that my PR is easy to review (not too big,
includes comments)
- ~I have made updated relevant documentation files (in project README,
docs/, etc)~
- ~I have added a decision log entry if the change affects the
architecture or changes a significant technology~
- [x] I have notified a reviewer

<!-- Please strike through and check off all items that do not apply
(rather than removing them) -->

---------

Co-authored-by: Jasper Moelker <jasper@voorhoede.nl>