refactor: Type datocmsRequest based on output mode or prerender. (#254)
In order to cut down on type assertion via `as`, `datocmsRequest` should
return a more workable type definition. Based on a passed type
variable – or when omitted, the value of `output` in the astro config –
`datocmsRequest` either has a return type that assumes the resource is
available, or passes the expected return value based on the query.

# Changes

- Export output mode from astro config.
- Allow asserted type override of datocmsRequest via `AssertReturnType`.

# How to test

1. Check out code locally
2. Open `@pages/[locale]/[...path]/index.astro`
3. Verify that the return value of `datocmsRequest` is typed by
`PageQuery`, without any type assertion.
4. Replace comment `/* typeof prerender */` with a `true` or `false`
value in the `datocmsRequest` type variables.
5. Verify that any mention of `page` gets TypeScript errors.
6. Remove the manipulated type varable.
7. See that the TypeScript errors disappear.
8. Change the output mode in `config/output.ts`;
5. Verify that the TypeScript errors return.

# Checklist

- [x] I have performed a self-review of my own code
- [x] I have made sure that my PR is easy to review (not too big,
includes comments)
- [ ] ~I have made updated relevant documentation files (in project
README, docs/, etc)~
- [ ] ~I have added a decision log entry if the change affects the
architecture or changes a significant technology~
- [x] I have notified a reviewer

<!-- Please strike through and check off all items that do not apply
(rather than removing them) -->