feat: more helpful errors when blocks are missing (#260)
Report on possible issues that can occur when a new block is not added
correctly.
# Changes

- Check if there are fields other than `__typename` are passed for a
specific block. If not, the block is probably missing in the query.
- Check if an Astro component is found for `__typename`. If not, the
component is probably not added to the `blocksByTypename` map.
- Better Typescript suppression
- More helpful hint
- Logging on server side when in development.


# How to test

1. Check out & run the project locally
2. Go to the query.graphql for the home page.
3. Remove the TextBlockRecord spread
4. Visit the homepage.
5. Verify that you see `No block data found for "TextBlockRecord". Is
the associated fragment included in the query?` in both your terminal
and the browser console.
6. Remove the `TextBlockRecord` entry in @blocks/Blocks.astro`
7. Visit the homepage.
8. Verify that you see `No block data found for "TextBlockRecord". Is it
included in <path to file>?` in both your terminal and the browser
console.

# Checklist

- [x] I have performed a self-review of my own code
- [x] I have made sure that my PR is easy to review (not too big,
includes comments)
- [ ] ~I have made updated relevant documentation files (in project
README, docs/, etc)~
- [ ] ~I have added a decision log entry if the change affects the
architecture or changes a significant technology~
- [x] I have notified a reviewer

<!-- Please strike through and check off all items that do not apply
(rather than removing them) -->