refactor: allow string type in setLocale; mock locales in tests (#264)
Adds type-guard for locale and updates setLocale and tests with
type-guard in place.
This allows tests relying on the locale to run with mock locales but
provides type guard functions that assert strings to be a locale when
passing.
Consider the following example in a function;
```ts
const locale = 'non-existent-locale';
if (!isLocale(locale)) return;
locales.includes(locale)
```
With `isLocale()`, the locale variable is now asserted to be a locale
when the check passes, without linting errors when using a non existent
locale.
Initially Typescript would assert locales.includes to contain an error
during linting, resulting in head-start projects test/lint CI to fail
when not all locales are present in the corresponding dato instance.
# Changes
- Adds `isLocale()` type-guard
- Updates `setLocale()` signature to accept any string
- Updates `datocmsSearch.test.ts` to map over mock locales that are cast
to SiteLocales as the test does not need to rely on existing locales


# How to test

1. CI lint/test

# Checklist

- [x] I have performed a self-review of my own code
- [x] I have made sure that my PR is easy to review (not too big,
includes comments)
- [x] ~I have made updated relevant documentation files (in project
README, docs/, etc)~
- [x] ~I have added a decision log entry if the change affects the
architecture or changes a significant technology~
- [x] I have notified a reviewer

<!-- Please strike through and check off all items that do not apply
(rather than removing them) -->