fix: set sameSite property to 'lax' to make cookie work when navigating from external site (#286)
# Changes

Set `sameSite: 'lax'` for the preview mode cookie, otherwise it will not
work when linked from a link on a page of a different domain. This is
currently happening when linking to the preview mode url from Dato,
where it always prompts for the secret, even when it's included in the
url.

# How to test

Enter the preview mode endpoint with the configured
([https://fix-preview-mode.head-start.pages.dev/api/preview/enter?secret=[secret]](https://fix-preview-mode.head-start.pages.dev/api/preview/enter?secret=[secret]))
secret from a link on a page and see if it directly goes into preview
mode instead of prompting for the secret.

# Checklist

- [ ] I have performed a self-review of my own code
- [x] I have made sure that my PR is easy to review (not too big,
includes comments)
- ~~I have made updated relevant documentation files (in project README,
docs/, etc)~~
- ~~I have added a decision log entry if the change affects the
architecture or changes a significant technology~~
- [x] I have notified a reviewer

<!-- Please strike through and check off all items that do not apply
(rather than removing them) -->